add ut for combined indexer & move to a seperate file
